SPMD: add support for an example SPM core manifest
This patch repurposes the TOS FW configuration file as the manifest for the SPM
core component which will reside at the secure EL adjacent to EL3. The SPM
dispatcher component will use the manifest to determine how the core component
must be initialised. Routines and data structure to parse the manifest have also
been added.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>