SPMD: [EXP] Groundwork for PSCI usage to initialise SPMC
This patch is the first in a series to demonstrate experimental support for use
of PSCI functions to initialise the SPMC on secondary CPUs.

This patch enables a SPMD to register a handler for PSCI calls invoked by the
SPMC. It also moves a couple of utility functions from a private PSCI header to
a public header. These functions will be used by the secure PSCI call handler.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>