SPMD: [EXP] Export entry point info as a global variable
This is a temporary solution, all secure world PSCI calls are meant to be
handled within S-EL2 which means the PSCI SPD callbacks need to be routed
to S-EL2 appropriately.

This patch converts the local entry_point_info structure used on the
boot cpu to a global as it will be used to initialise the SPMC context
on secondary CPUs in future patches.

Change-Id: Icb92a2403d746cf120f05708bab56d3ffb3f7e05
Signed-off-by: Achin Gupta <achin.gupta@arm.com>