SPMD: [EXP] Use PSCI states to manage SPMC boot
This is a temporary solution, all secure world PSCI calls are meant to be
handled within S-EL2 which means the PSCI SPD callbacks need to be routed
to S-EL2 appropriately.

This patch replaces private states used by the SPMD to determine whether
SPMC has booted on a CPU with PSCI affinity states. This enables the SPMC
on the boot CPU to use the PSCI_CPU_ON function to begin initialisation
on secondary CPUs.

Change-Id: Ie29b31c18f1e9ede199dc2c9c506fafb13fed296
Signed-off-by: Achin Gupta <achin.gupta@arm.com>