SPMD: [EXP] Add PSCI_CPU_ON support to SPMD
This is a temporary solution, all secure world PSCI calls are meant to be
handled within S-EL2 which means the PSCI SPD callbacks need to be routed
to S-EL2 appropriately.

This patch enables the SPMD and SPMC to use the PSCI_CPU_ON function
to initialise SPMC contexts on secondary CPUs. It registers a handler
for PSCI calls from the secure world. Only the PSCI_CPU_ON call is
implemented. It is used to register the entry point for the SPMC on
each secondary CPU.

It also registers a handler for a PSCI CPU ON finish operation that was
invoked by the normal world. This handler initialises the SPMC context
on the secondary CPU in response to an invocation of this hook.

Change-Id: I4412b9a97afb520aaf9931d1c82979c2831c34b9
Signed-off-by: Achin Gupta <achin.gupta@arm.com>