SPMD: [EXP] Add PSCI_CPU_ON support to SPMD
This patch enables the SPMD and SPMC to use the PSCI_CPU_ON function to
initialise SPMC contexts on secondary CPUs. It registers a handler for PSCI
calls from the secure world. Only the PSCI_CPU_ON call is implemented. It is
used to register the entry point for the SPMC on each secondary CPU.

It also registers a handler for a PSCI CPU ON finish operation that was invoked
by the normal world. This handler initialises the SPMC context on the secondary
CPU in response to an invocation of this hook.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>