Add build script and README for prototype
Adds a PROTOTYPE_README detailing the required environment and a
`multi_build.sh` script to build/run the prototype.

The script is able to run the prototype both with and without EL-2
enabled in the normal and secure world independently. For more
usage information please consult the README or invoke the script
with `--help`.

Change-Id: Id489ad943399ff1da8976553f9d1318b3be6e4b5