qemu: use sbsa-ec for reboot/poweroff secure machine
qemu just merdge sbsa-ec driver which is simple memory
mapped power controler for emulated device. Adopting
this driver to machine virt,secure=on makes it possible
to follow common firmware way for reboot and poweroff.
I.e. in secure boot PSCI control moves to firmware and
after that we need some way say qemu that we are ready
for reboot.

Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>