refactor(spm): rename common data structures and functions
Subsequent to the restructuring of the spm_mm code, the names of
the data structures and functions that are now common to other
variants of secure partition manager (such as the EL3 SPMC) are renamed.
The 'mm' in the 'spm_mm' part of the name is removed to indicate that
those data structures and functions are usable for multiple secure
partition manager implementations.

Signed-off-by: Sayanta Pattanayak <sayanta.pattanayak@arm.com>
Change-Id: Iecad184ecf52f6fdabf276c1acfad25f75750d80