STMM: Temporarily move STMM code to LP
This patch moves the STMM specific code into
a separate STMM logical partition file. The future
aim is to add support for the required functionality
direct to the SPMC in the form of FFA ABIs instead
of requiring this functionality to be separated.

Refactor the left behind code to be more generic and
suitable for non STMM partitions as well.

Change-Id: I62030d2afee3fcf91972fdfb8767949689e10688
Signed-off-by: Marc Bonnici <marc.bonnici@arm.com>