REVISIT: Separate StMM SP specifics to add support for a S-EL1 SP
This patch separates code common to both the MM communication interface SPM and
FF-A EL3 SPMC to support an StMM SP in S-EL0. This allows to better understand
what is required to support a S-EL1 SP instead of trying to retrofit what exists
to make it work for both S-EL0 and S-EL1 SPs. This patch needs to be revisited
to understand if the separation needs to formalised or reconsidered.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>