feat(spmc): Implement FF-A Notifications
Provide SP -> VM notifications as per FF-A specification v1.1 EAC0.
This module could be adapted to also support VM -> SP and Framework
notifications.

FF-A Notifications are an asynchronous communication mechanism
designed for reliability.  Since its primitives are merely abstrac-
tions of shared state, the mechanism is not subject to capacity
limits in the same way messaging is, but on its own it is not suit-
able for conveying arbitrary data.

Change-Id: I7c69eb662c9f99a9c885aa88dc641640a6e21e83